Extract Homer face
