+++ agoodWatch/homerFace.h
@@ -1,0 +1,17 @@
+/*
+ * Copyright (c) 2020 Alex Goodyear
+ * Derived from ...
+Copyright (c) 2019 lewis he
+This is just a demonstration. Most of the functions are not implemented.
+The main implementation is low-power standby. 
+The off-screen standby (not deep sleep) current is about 4mA.
+Select standard motherboard and standard backplane for testing.
+Created by Lewis he on October 10, 2019.
+*/
+
+#ifndef __HOMER_FACE_H
+#define __HOMER_FACE_H
+
+int createHomerFace (int);
+
+#endif /*__HOMER_FACE_H */
